Add Project Member functionality and project invitations

+++ src/main/java/com/example/sprintsight/services/ProjectService.java
@@ -5,49 +5,50 @@ import com.example.sprintsight.dtos.requests.UpdateProjectRequest;
 import com.example.sprintsight.dtos.responses.ProjectResponse;
 import com.example.sprintsight.entities.Project;
 import com.example.sprintsight.entities.User;
-import com.example.sprintsight.exceptions.EntityNotFoundException;
 import com.example.sprintsight.mappers.ProjectMapper;
 import com.example.sprintsight.repositories.ProjectRepository;
-import com.example.sprintsight.repositories.UserRepository;
-import com.example.sprintsight.security.UserPrincipal;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class ProjectService {
-    private final ProjectRepository projectRepository;
-    private final UserRepository userRepository;
+    private final UserService userService;
     private final ProjectMapper projectMapper;
+    private final ProjectRepository projectRepository;
 
+    @Transactional(readOnly = true)
     public ProjectResponse getProject(UUID id) {
         return projectMapper.toProjectResponse(findProject(id));
     }
 
-    public List<ProjectResponse> getProjects() {
-        return projectRepository.findAll()
+    @Transactional(readOnly = true)
+    public List<ProjectResponse> getOwnedProjects(UUID userId) {
+        return projectRepository.findByCreatedBy_Id(userId)
+                .stream()
+                .map(projectMapper::toProjectResponse)
+                .toList();
+    }
+
+    @Transactional(readOnly = true)
+    public List<ProjectResponse> getMemberProjects(UUID userId) {
+        return projectRepository.findAllProjectsByMemberId(userId)
                 .stream()
                 .map(projectMapper::toProjectResponse)
                 .toList();
     }
 
     @Transactional
-    public ProjectResponse addProject(CreateProjectRequest request) {
-        UserPrincipal principal = (UserPrincipal) Objects.requireNonNull(
-                SecurityContextHolder.getContext().getAuthentication()
-        ).getPrincipal();
-
-        assert principal != null;
-        User user = userRepository.findById(principal.getId())
-                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+    public ProjectResponse addProject(CreateProjectRequest request, UUID principalId) {
+        User user = userService.findUser(principalId);
 
         Project project = projectMapper.toEntity(request);
         project.setCreatedBy(user);
@@ -56,8 +57,15 @@ public class ProjectService {
     }
 
     @Transactional
-    public ProjectResponse updateProject(UpdateProjectRequest request, UUID id, boolean isPut) {
-        Project project = findProject(id);
+    public ProjectResponse updateProject(
+            UpdateProjectRequest request,
+            UUID projectId,
+            UUID principalId,
+            boolean isPut
+    ) {
+        Project project = findProject(projectId);
+
+        verifyOwnership(project, principalId);
 
         if (isPut) {
             projectMapper.updateProjectFromPut(request, project);
@@ -70,16 +78,24 @@ public class ProjectService {
     }
 
     @Transactional
-    public void deleteProject(UUID id) {
-        Project project = findProject(id);
+    public void deleteProject(UUID projectId, UUID principalId) {
+        Project project = findProject(projectId);
+
+        verifyOwnership(project, principalId);
 
         projectRepository.delete(project);
 
-        log.info("Deleted project: {}", id);
+        log.info("Deleted project: {}", projectId);
     }
 
-    private Project findProject(UUID id) {
+    protected Project findProject(UUID id) {
         return projectRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Project not found"));
+    }
+
+    private void verifyOwnership(Project project, UUID principalId) {
+        if (!project.getCreatedBy().getId().equals(principalId)) {
+            throw new AccessDeniedException("You do not have permission to modify this project");
+        }
     }
 
     private ProjectResponse saveProject(Project project, String logMessage) {
